+++ tests/edges/test_arc2d.py
@@ -105,11 +105,11 @@ class TestArc2D(unittest.TestCase):
         for result_list, expected_result_list in zip(list_point_belongs, expected_results):
             self.assertEqual(result_list, expected_result_list)
 
-        arc = Arc2D.load_from_file(os.path.join(folder, "arc2d_bug_point_belongs.json"))
+        arc = Arc2D.from_json(os.path.join(folder, "arc2d_bug_point_belongs.json"))
         point = volmdlr.Point2D(0.01330629098214331, 0.0032923224261096617)
         self.assertTrue(arc.point_belongs(point))
 
-        arc = Arc2D.load_from_file(os.path.join(folder, "arc2d_point_belongs.json"))
+        arc = Arc2D.from_json(os.path.join(folder, "arc2d_point_belongs.json"))
         point = volmdlr.Point2D(0.0007151488183559929, 0.007258543823331798)
         self.assertTrue(arc.point_belongs(point))
 
@@ -177,7 +177,7 @@ class TestArc2D(unittest.TestCase):
         rotated_arc2d = rotated_arc2d.rotation(volmdlr.O2D, -math.pi / 2)
         self.assertTrue(rotated_arc2d.is_close(self.arc2))
 
-        arc2d = Arc2D.load_from_file(os.path.join(folder, "arc2d_rotation_test.json"))
+        arc2d = Arc2D.from_json(os.path.join(folder, "arc2d_rotation_test.json"))
         rotated_arc2d = arc2d.rotation(volmdlr.Point2D(0.5, 0.5), math.pi / 1.5)
         self.assertEqual(arc2d.frame.u, rotated_arc2d.frame.u)
         self.assertEqual(arc2d.frame.v, rotated_arc2d.frame.v)