+++ tests/edges/test_arcellipse2d.py
@@ -133,7 +133,7 @@ class TestArcEllipse2D(unittest.TestCase):
         for point, expected_point in zip(list_points, expected_points):
             self.assertTrue(point.is_close(expected_point))
 
-        arcellipse = edges.ArcEllipse2D.load_from_file(os.path.join(folder, "ellipse2d_point_at_abscissa.json"))
+        arcellipse = edges.ArcEllipse2D.from_json(os.path.join(folder, "ellipse2d_point_at_abscissa.json"))
         abscissa = 1.4594044224379008
         point = arcellipse.point_at_abscissa(abscissa)
         self.assertTrue(point.is_close(volmdlr.Point2D(0.2409700344646443, -0.10841585996396141)))