+++ CHANGELOG.md
@@ -43,6 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - load_from_file -> from_json
 - Delete outdated_methods: bbox_intersections -> is_intersecting | b_rectangle_intersections -> is_intersecting
+- For points in the interior of an object: point_belongs -> point_inside
+- For points ON an object: point_belongs
 - edges/curves.py cut_between_two_points -> trim
 - defines ordering of curve methods